requirements.txt
langchain==0.1.0
langchain-google-vertexai>=0.0.4
google-cloud-aiplatform>=1.26.1
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=1.10.7
python-multipart>=0.0.6
